Added entryStream helper.

+++ src/main/scala/com/yogpc/qp/tile/TileAdvPump.scala
@@ -257,7 +257,7 @@ class TileAdvPump extends APowerTile with IEnchantableTile with ITickable with I
     /**
       * @return Map (Enchantment id, level)
       */
-    override def getEnchantments = ench.getMap.collect { case (a, b) if b > 0 => (Int.box(a), Byte.box(b)) }.asJava
+    override def getEnchantments = ench.getMap.collect { case (a, b) if b > 0 => (Int.box(a), Int.box(b)) }.asJava
 
     /**
       * @param id    Enchantment id
